Committing before merge.

+++ OPiT/PnPSolver.h
@@ -14,8 +14,8 @@ public:
 	std::vector<cv::Point3d> getWorldPoints();
 
 	void setImagePoints();
-	void setImagePoints(std::vector<cv::Point2f> IP);
-	std::vector<cv::Point2f> getImagePoints();
+	void setImagePoints(std::vector<cv::Point2d> IP);
+	std::vector<cv::Point2d> getImagePoints();
 
 	cv::Mat getCameraPose();
 	cv::Mat getCameraPosition();
@@ -41,7 +41,7 @@ public:
 
 private:
 	
-	std::vector<cv::Point2f> imagePoints, imagePoints2;
+	std::vector<cv::Point2d> imagePoints, imagePoints2;
 	std::vector<cv::Point3d> worldPoints;
 	std::vector<std::vector<cv::Point2d> > VoVImagePoints;
